+++ src/test/java/soon/springtestutil/example/TeamRepository.java
@@ -1,0 +1,7 @@
+package soon.springtestutil.example;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamRepository extends JpaRepository<Team, Long> {
+
+}